requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.7
iniconfig==1.1.1
packaging==21.0
Pillow==8.3.2
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytz==2021.1
qrcode==7.3
sqlparse==0.4.2
toml==0.10.2
uuid==1.30
